Split up state tester
Removed a lot of stuff and split the StateTester file into States, StateManager, and StateTester
