--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5d1a22a8a18cdec488b6a66c5314d523
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
