--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Attacking Gate.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f3852870910d7c247827877c3ee06ec6
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
