--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Battling.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f14d60fed6d321b409a95946087bbad1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
