--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Choosing Gate.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: da91e3d4e8039094d865288f3bdd5f4c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
