--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Rolling Gate Effect.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bfeb732d82f05ee41b830fda7fae8a47
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
