--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Rolling Trait.asset
@@ -1,15 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 76893895183749f4fbfa1b4389388a53, type: 3}
-  m_Name: Rolling Trait
-  m_EditorClassIdentifier: 
-  gameState: 0
