--- GatekeeperStudyHall/Assets/ScriptableObjects/GameStateInstances/Rolling Trait.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2694a438d7028e0439caad35f5cb5d37
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
