+++ GatekeeperStudyHall/Assets/Scripts/StateMachine.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 76893895183749f4fbfa1b4389388a53
+guid: 6d3b295ef571cf940b16e7feb2de4ce5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
